Special emergency patch for "stealing first" play.
With this patch, events in the game between the Cubs and Brewers
on 2013-04-19 can be accommodated on a provisional basis.

In that game, Jean Segura "stole" first base after having attained
second base safely.

This patch permits the "advance" of a runner from second to first. The
sequence in question can be coded:

play,8,1,430001,12,BCC,OA.1X2(15)
play,8,1,430001,12,BCC.,OA.2-1
play,8,1,430001,12,BCC..*S,K
play,8,1,518960,01,C,CS2(24)

The two OAs were actually one play, but this patch does not support
the crossing advancements.

This is intended only for emergency use until a pronouncement is made
as to the way Retrosheet will code this play.  This is not recommended
as a permanent solution, as it permits backwards advancement without
restriction - and in almost all instances, that should be flagged
as an error.